Codigo/massa_backend/requirements/requirements.txt
fastapi[standard]==0.115.0
pydantic-settings==2.6.0
asyncpg==0.30.0
SQLAlchemy==2.0.36
rdkit==2024.3.5
numpy==1.26.4
pandas==2.2.3
matplotlib==3.9.2
scipy==1.14.1
scikit-learn==1.5.2
kmodes==0.12.2
PyJWT==2.10.0
reportlab==4.2.5
passlib==1.7.4
bcrypt==4.2.1